:recycle: separate text size tests
